# src/api/openapi.yaml
openapi: 3.1.1
info:
  title: SDL API
  version: 2.5.0-dev
  description: Stream, download, inspect, and resolve supported SDL video links.
servers:
  - url: http://localhost:8080
paths:
...
  /api/download:
    get:
      operationId: downloadFromQuery
      summary: Download a resolved video as an attachment
      parameters:
        - $ref: '#/components/parameters/Url'
        - $ref: '#/components/parameters/Language'
        - $ref: '#/components/parameters/VideoType'
        - $ref: '#/components/parameters/Episodes'
        - $ref: '#/components/parameters/Seasons'
        - $ref: '#/components/parameters/Extractor'
        - $ref: '#/components/parameters/UserAgent'
        - $ref: '#/components/parameters/Referer'
      responses:
        '200':
          description: Downloadable video response
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            video/mp4:
              schema:
                type: string
                format: binary
    post:
      operationId: download
      summary: Download a resolved video as an attachment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayRequest'
      responses:
        '200':
          description: Downloadable video response
...
components:
  parameters:
    Url:
      name: url
      in: query
      required: true
      schema:
        type: string
        format: uri
    Language:
      name: language
      in: query
      schema:
        type: string
        enum: [unspecified, any, german, english]
    VideoType:
      name: video_type
      in: query
      schema:
        type: string
        enum: [unspecified, any, raw, dub, sub]
    Episodes:
      name: episodes
      in: query
      schema:
        type: string
        description: Single episode, comma/range list, or all.
    Seasons:
      name: seasons
      in: query
      schema:
        type: string
        description: Single season, comma/range list, or all.
    Extractor:
      name: extractor
      in: query
      schema:
        type: string
    UserAgent:
      name: user_agent
      in: query
      schema:
        type: string
    Referer:
      name: referer
      in: query
      schema:
        type: string
        format: uri
  schemas:
    PlayRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
        language:
          type: string
        video_type:
          type: string
        episodes:
          type: string
        seasons:
          type: string
        extractor_priorities:
          type: array
          items:
            type: string
        extractor:
          type: string
        user_agent:
          type: string
        referer:
          type: string
          format: uri
...